Fixed a very interesting bug about not going to the next track.
This bug was first caused by not correctly checking if the music is still playing. After fixing that, the program would crash because in the very first second, media_player.is_playing would return 0 resulting in program crash. I fixed it by adding an playtime checker so the test would run correctly.